move cecdu into 3ds/services/ folder
