+++ Makefile
@@ -44,6 +44,9 @@ GFXBUILD	:=	$(BUILD)
 #ROMFS		:=	romfs
 #GFXBUILD	:=	$(ROMFS)/gfx
 
+SOURCES		+= source/3ds/services
+INCLUDES	+= include/3ds/services
+
 #---------------------------------------------------------------------------------
 # options for code generation
 #---------------------------------------------------------------------------------
