+++ include/Box.hpp
@@ -5,9 +5,10 @@
 #include "OBIndex.hpp"
 #include <stdio.h>
 #include <string>
+
 extern "C"
 {
-#include "cecdu.h"
+#include "3ds/services/cecdu.h"
 }
 
 class Box
