+++ source/BoxInfo.cpp
@@ -1,6 +1,7 @@
 #include "BoxInfo.hpp"
+
 extern "C" {
-#include "cecdu.h"
+#include "3ds/services/cecdu.h"
 #include <string.h>
 }
